pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "docmark"
version = "0.1.0"
description = "A library for converting between Word files and Markdown."
readme = "README.md"
requires-python = ">=3.8"
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "click",
    "mammoth",
    "markdownify",
    "markdown",
    "python-docx",
    "beautifulsoup4",
    "Pillow",
    "requests",
    "exifread",
]

[project.optional-dependencies]
llm = ["openai", "anthropic"]
pdf = ["pdfminer.six", "pypdf", "pdfplumber", "pdf2image", "pytesseract"]
all = ["docmark[llm,pdf]"]

[project.scripts]
docmark = "docmark.cli:main"

[tool.setuptools.packages.find]
where = ["."]  # Look for packages in the current directory